source files now use arguments passed in by the user
when sourcing a file, the parent function's arguments are used if nothing is passed into source.

this creates confusing results, as the function in this case is the tome
wrapper function, which includes arguments such as the function
name itself.

Including an empty argument as part of the source invocation overrides
arguments.

using . vs source, as "." is the posix-standard way of sourcing a file.