ignore directory support, alphabetical commands
Both listing of directories and completion are sorted alphabetically now

It is now possible to add a .tomeignore file to ignore a directory from
the command list.

fixed a bug where the help command did not work when quotes
were included.

support for usage now exists. USAGE prefixes are now what appears
with the help command, rather than the help docs.