bugfix: /bin/bash is now a recognized shell
Common shells like /bin/bash were no longer recognized as a shell,
which is problematic as /bin/bash is a valid value for $0, which is
documented to be a valid initialization process for tome.

Generally tome was not recognizing shells that had a path. This ensures
that the file stem is used to match the shell.

Fixes #17