reimplement CLI using clap (#23)
The original CLI parsing code was very rudimentary, so there
was a need to replace it with something library.

Some changes that occurred due to that:

The underlying commands of tome itself were renamed to align more
with the design proposal. One deviation was prefixing things that
were supposed to be used by tome instances with "command", to clarify
that those are expected to be run by the instance instead.

Some minor refactoring to clarify the command handlers.