Require "COMPLETION" flag to complete script
tome by default attempted to get completion for a script by executing
it.

As the default behavior, this can be error-prone: the user may not
want their script to execute because it can cause destructive behavior,
and that could be executed with a tab button pressed via muscle memory.

If completion is supported, require the user to explicitly state as
such instead.