feat: Add `tome run`
tome run satisfies a use case where one would like to re-use
tome's script navigation functionality, but in an environment
that is not an interactive shell.

This could be useful in a situation where one would like to
share scripts across a local and deployed envioronment.

tome run does an execve to effectively become the function
invoked: although this is less cross-platform compatible, it
results in the simplest method to reduce complications like
buffer sizes handing off data in intermediary pipes.

refactoring some finder logic out for re-use.

fixes #40.