fix: fix completion for sh/bash/zsh (#47)
Issue #43 reported compgen failing on completion parameters
that looked like they were command line switches (e.g. "--")

This looks to be due to incorrect usage of compgen, where the difference
between options and the user's completion  to complete were not separated with
the "--" delimiter. According to getopt(1) documentation this should be
split explicitly:

       When the end of options is encountered, the getopts utility shall
       exit with a return value greater than zero; the shell variable
       OPTIND shall be set to the index of the first operand, or the
       value "$#"+1 if there are no operands; the name variable shall be
       set to the <question-mark> character. Any of the following shall
       identify the end of options: the first "--" argument that is not
       an option-argument, finding an argument that is not an option-
       argument and does not begin with a '-', or encountering an error.

This fixes the issue when reproing locally.